Update sources
* clean up document shelf
* remove semantgic information
  from options
* enable multiple records to
  match further Submodel
  templates
* "Save as .." for digital file
  and drag from image
* display preview image
