+++ src/AasxPluginDocumentShelf/DocumentEntity.cs
@@ -9,10 +9,14 @@ This source code may use other Open Source software components (see LICENSE.txt)
 
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
 using System.Windows.Markup;
+using System.Windows.Media.Imaging;
 using AasxIntegrationBase;
 using AasxPredefinedConcepts;
 using AdminShellNS;
@@ -30,7 +34,9 @@ namespace AasxPluginDocumentShelf
         public delegate void MenuClickDelegate(DocumentEntity e, string menuItemHeader, object tag);
         public event MenuClickDelegate MenuClick = null;
 
-        public enum SubmodelVersion { Default, V11 }
+        public event DocumentEntityEvent DragStart = null;
+
+        public enum SubmodelVersion { Default, V10, V11 }
 
         public SubmodelVersion SmVersion = SubmodelVersion.Default;
 
@@ -38,8 +44,7 @@ namespace AasxPluginDocumentShelf
         public string Organization = "";
         public string FurtherInfo = "";
         public string[] CountryCodes;
-        public string DigitalFile = "";
-        public string MimeType = null;
+        public FileInfo DigitalFile, PreviewFile;
         public System.Windows.Controls.Viewbox ImgContainer = null;
         public string ReferableHash = null;
 
@@ -52,6 +57,20 @@ namespace AasxPluginDocumentShelf
         public enum DocRelationType { DocumentedEntity, RefersTo, BasedOn, Affecting, TranslationOf };
         public List<Tuple<DocRelationType, AdminShell.Reference>> Relations =
             new List<Tuple<DocRelationType, AdminShellV20.Reference>>();
+
+        public class FileInfo
+        {
+            public string Path = "";
+            public string MimeType = "";
+
+            public FileInfo() { }
+
+            public FileInfo(AdminShell.File file)
+            {
+                Path = file?.value;
+                MimeType = file?.mimeType;
+            }
+        }
 
         public DocumentEntity() { }
 
@@ -71,13 +90,43 @@ namespace AasxPluginDocumentShelf
 
         public void RaiseMenuClick(string menuItemHeader, object tag)
         {
-            if (MenuClick != null)
-                MenuClick(this, menuItemHeader, tag);
+            MenuClick?.Invoke(this, menuItemHeader, tag);
+        }
+
+        public void RaiseDragStart()
+        {
+            DragStart?.Invoke(this);
+        }
+
+        /// <summary>
+        /// This function needs to be called a part of tick-Thread in STA / UI thread
+        /// </summary>
+        public BitmapImage LoadImageFromPath(string fn)
+        {
+            // be a bit suspicous ..
+            if (!File.Exists(fn))
+                return null;
+
+            // convert here, as the tick-Thread in STA / UI thread
+            try
+            {
+                var bi = new BitmapImage(new Uri(fn, UriKind.RelativeOrAbsolute));
+                var img = new Image();
+                img.Source = bi;
+                ImgContainer.Child = img;
+                return bi;
+            } catch (Exception ex)
+            {
+                LogInternally.That.SilentlyIgnoredError(ex);
+            }
+            return null;
         }
     }
 
     public class ListOfDocumentEntity : List<DocumentEntity>
     {
+        private static DocuShelfSemanticConfig _semConfig = DocuShelfSemanticConfig.Singleton;
+
         //
         // Default
         //
@@ -95,7 +144,7 @@ namespace AasxPluginDocumentShelf
             if (subModel?.submodelElements != null)
                 foreach (var smcDoc in
                     subModel.submodelElements.FindAllSemanticIdAs<AdminShell.SubmodelElementCollection>(
-                        options?.SemIdDocument, AdminShellV20.Key.MatchMode.Relaxed))
+                        _semConfig.SemIdDocument, AdminShellV20.Key.MatchMode.Relaxed))
                 {
                     // access
                     if (smcDoc == null || smcDoc.value == null)
@@ -104,7 +153,7 @@ namespace AasxPluginDocumentShelf
                     // look immediately for DocumentVersion, as only with this there is a valid List item
                     foreach (var smcVer in
                         smcDoc.value.FindAllSemanticIdAs<AdminShell.SubmodelElementCollection>(
-                            options?.SemIdDocumentVersion, AdminShellV20.Key.MatchMode.Relaxed))
+                            _semConfig.SemIdDocumentVersion, AdminShellV20.Key.MatchMode.Relaxed))
                     {
                         // access
                         if (smcVer == null || smcVer.value == null)
@@ -117,13 +166,13 @@ namespace AasxPluginDocumentShelf
                         // take the 1st title
                         var title =
                             "" +
-                            smcVer.value.FindFirstSemanticIdAs<AdminShell.Property>(options?.SemIdTitle,
+                            smcVer.value.FindFirstSemanticIdAs<AdminShell.Property>(_semConfig.SemIdTitle,
                             AdminShellV20.Key.MatchMode.Relaxed)?.value;
 
                         // could be also a multi-language title
                         foreach (var mlp in
                             smcVer.value.FindAllSemanticIdAs<AdminShell.MultiLanguageProperty>(
-                                options?.SemIdTitle, AdminShellV20.Key.MatchMode.Relaxed))
+                                _semConfig.SemIdTitle, AdminShellV20.Key.MatchMode.Relaxed))
                             if (mlp.value != null)
                                 title = mlp.value.GetDefaultStr(defaultLang);
 
@@ -131,24 +180,24 @@ namespace AasxPluginDocumentShelf
                         var orga =
                             "" +
                             smcVer.value.FindFirstSemanticIdAs<AdminShell.Property>(
-                                options?.SemIdOrganizationOfficialName, AdminShellV20.Key.MatchMode.Relaxed)?.value;
+                                _semConfig.SemIdOrganizationOfficialName, AdminShellV20.Key.MatchMode.Relaxed)?.value;
                         if (orga.Trim().Length < 1)
                             orga =
                                 "" +
                                 smcVer.value.FindFirstSemanticIdAs<AdminShell.Property>(
-                                    options?.SemIdOrganizationName, AdminShellV20.Key.MatchMode.Relaxed)?.value;
+                                    _semConfig.SemIdOrganizationName, AdminShellV20.Key.MatchMode.Relaxed)?.value;
 
                         // class infos
                         var classId =
                             "" +
                             smcDoc.value.FindFirstSemanticIdAs<AdminShell.Property>(
-                                options?.SemIdDocumentClassId, AdminShellV20.Key.MatchMode.Relaxed)?.value;
+                                _semConfig.SemIdDocumentClassId, AdminShellV20.Key.MatchMode.Relaxed)?.value;
 
                         // collect country codes
                         var countryCodesStr = new List<string>();
                         var countryCodesEnum = new List<AasxLanguageHelper.LangEnum>();
                         foreach (var cclp in
-                            smcVer.value.FindAllSemanticIdAs<AdminShell.Property>(options?.SemIdLanguage,
+                            smcVer.value.FindAllSemanticIdAs<AdminShell.Property>(_semConfig.SemIdLanguage,
                             AdminShellV20.Key.MatchMode.Relaxed))
                         {
                             // language code
@@ -188,10 +237,10 @@ namespace AasxPluginDocumentShelf
                         var further = "";
                         foreach (var fi in
                             smcVer.value.FindAllSemanticIdAs<AdminShell.Property>(
-                                options?.SemIdDocumentVersionIdValue))
+                                _semConfig.SemIdDocumentVersionIdValue))
                             further += "\u00b7 version: " + fi.value;
                         foreach (var fi in
-                            smcVer.value.FindAllSemanticIdAs<AdminShell.Property>(options?.SemIdDate,
+                            smcVer.value.FindAllSemanticIdAs<AdminShell.Property>(_semConfig.SemIdDate,
                             AdminShellV20.Key.MatchMode.Relaxed))
                             further += "\u00b7 date: " + fi.value;
                         if (further.Length > 0)
@@ -208,9 +257,9 @@ namespace AasxPluginDocumentShelf
 
                         // filename
                         var fl = smcVer.value.FindFirstSemanticIdAs<AdminShell.File>(
-                            options?.SemIdDigitalFile, AdminShellV20.Key.MatchMode.Relaxed);
-                        ent.DigitalFile = fl?.value;
-                        ent.MimeType = fl?.mimeType;
+                            _semConfig.SemIdDigitalFile, AdminShellV20.Key.MatchMode.Relaxed);
+
+                        ent.DigitalFile = new DocumentEntity.FileInfo(fl);
 
                         // add
                         ent.SmVersion = DocumentEntity.SubmodelVersion.Default;
@@ -226,7 +275,7 @@ namespace AasxPluginDocumentShelf
         // V11
         //
 
-        private static void SeachForRelations(
+        private static void SearchForRelations(
             AdminShell.SubmodelElementWrapperCollection smwc,
             DocumentEntity.DocRelationType drt,
             AdminShell.Reference semId,
@@ -377,8 +426,12 @@ namespace AasxPluginDocumentShelf
                         var further = "";
                         foreach (var fi in
                             smcVer.value.FindAllSemanticIdAs<AdminShell.Property>(
-                                defs11.CD_DocumentIdValue?.GetReference()))
+                                defs11.CD_DocumentVersionId?.GetReference()))
                             further += "\u00b7 version: " + fi.value;
+                        foreach (var fi in
+                            smcVer.value.FindAllSemanticIdAs<AdminShell.Property>(
+                                defs11.CD_DocumentIdValue?.GetReference()))
+                            further += "\u00b7 id: " + fi.value;
                         foreach (var fi in
                             smcVer.value.FindAllSemanticIdAs<AdminShell.Property>(defs11.CD_SetDate?.GetReference(),
                             AdminShellV20.Key.MatchMode.Relaxed))
@@ -395,20 +448,25 @@ namespace AasxPluginDocumentShelf
                         ent.SourceElementsDocument = smcDoc.value;
                         ent.SourceElementsDocumentVersion = smcVer.value;
 
-                        // filename
+                        // file informations
                         var fl = smcVer.value.FindFirstSemanticIdAs<AdminShell.File>(
-                            defs11.CD_DigitalFile?.GetReference());
-                        ent.DigitalFile = fl?.value;
-                        ent.MimeType = fl?.mimeType;
+                            defs11.CD_DigitalFile?.GetReference(), AdminShellV20.Key.MatchMode.Relaxed);
+                        if (fl != null)
+                            ent.DigitalFile = new DocumentEntity.FileInfo(fl);
+
+                        fl = smcVer.value.FindFirstSemanticIdAs<AdminShell.File>(
+                            defs11.CD_PreviewFile?.GetReference(), AdminShellV20.Key.MatchMode.Relaxed);
+                        if (fl != null)
+                            ent.PreviewFile = new DocumentEntity.FileInfo(fl);
 
                         // relations
-                        SeachForRelations(smcVer.value, DocumentEntity.DocRelationType.DocumentedEntity,
+                        SearchForRelations(smcVer.value, DocumentEntity.DocRelationType.DocumentedEntity,
                             defs11.CD_DocumentedEntity?.GetReference(), ent);
-                        SeachForRelations(smcVer.value, DocumentEntity.DocRelationType.RefersTo,
+                        SearchForRelations(smcVer.value, DocumentEntity.DocRelationType.RefersTo,
                             defs11.CD_RefersTo?.GetReference(), ent);
-                        SeachForRelations(smcVer.value, DocumentEntity.DocRelationType.BasedOn,
+                        SearchForRelations(smcVer.value, DocumentEntity.DocRelationType.BasedOn,
                             defs11.CD_BasedOn?.GetReference(), ent);
-                        SeachForRelations(smcVer.value, DocumentEntity.DocRelationType.TranslationOf,
+                        SearchForRelations(smcVer.value, DocumentEntity.DocRelationType.TranslationOf,
                             defs11.CD_TranslationOf?.GetReference(), ent);
 
                         // add
